requirements.txt
fastapi==0.115.5
huggingface-hub==0.26.2
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.18
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2
langsmith==0.1.143
PyMuPDF==1.24.13
python-dotenv==1.0.1
SQLAlchemy==2.0.35
starlette==0.41.2
transformers==4.46.2
uvicorn==0.32.0
psycopg2==2.9.10
python-multipart==0.0.17
faiss-cpu==1.7.3
